added text search components
