+++ src/app/components/search-results/search-result.component.html
@@ -18,23 +18,7 @@
         <div class="col-xs-12" *ngIf="showShuffled && imagesSearch">
           <app-image-results [data]="shuffledSearchResults"></app-image-results>
         </div>
-
         
-        <div class="col-xs-6" *ngIf="showShuffled && !imagesSearch">
-          <ng-container *ngIf="!imagesSearch; else imagesContainerShuff">
-            <div class="results-container">
-              <app-text-results [data]="shuffledSearchResults"></app-text-results>
-            </div>
-          </ng-container>
-        </div>
-        <div class="col-xs-6" *ngIf="(isMobile && !showShuffled) || (!isMobile) || !imagesSearch">
-          
-          <ng-container *ngIf="!imagesSearch; else imagesContainerUnShuff">
-            <div class="results-container">
-              <app-text-results [data]="unshuffledSearchResults"></app-text-results>
-            </div>
-          </ng-container>
-        </div>
       </div>
     </div>
   </ng-container>
