+++ src/app/components/search-results/text-results/text-list/text-item.type.ts
@@ -1,0 +1,3 @@
+export interface textItem {
+    
+}
