+++ src/app/components/search-results/text-results/text-list/text-list.component.html
@@ -1,0 +1,3 @@
+<div class="results-container">
+    <app-text-row *ngFor="let index of data" [data]="index"></app-text-row>
+</div>
