+++ src/app/components/search-results/text-results/text-list/text-list.component.ts
@@ -1,0 +1,18 @@
+import { Component, OnInit, Input } from '@angular/core';
+import { textItem } from './text-item.type';
+
+@Component({
+  selector: 'app-text-list',
+  templateUrl: './text-list.component.html',
+  styleUrls: ['./text-list.component.css']
+})
+export class TextListComponent implements OnInit {
+
+    @Input() data: textItem[];
+
+    constructor() { }
+
+    ngOnInit() {
+    }
+
+}
