+++ src/app/components/search-results/text-results/text-list/text-row/text-item.component.html
@@ -1,0 +1,8 @@
+<div class="result-div">
+    <div class="original-position">
+      Position: {{ data.originaldataultIndex + 1 }}
+    </div>
+    <a href="{{ data.url }}" class="title">{{ data.name }}</a>
+    <a href="{{ data.url }}" class="url">{{ data.url }}</a>
+    <div class="content">{{ data.snippet }}</div>
+</div>
