+++ src/app/components/search-results/text-results/text-results.component.html
@@ -1,10 +1,19 @@
-<div class="results-container">
-  <div class="result-div" *ngFor="let res of data; let i = index">
-    <div class="original-position">
-      Position: {{ res.originalResultIndex + 1 }}
-    </div>
-    <a href="{{ res.url }}" class="title">{{ res.name }}</a>
-    <a href="{{ res.url }}" class="url">{{ res.url }}</a>
-    <div class="content">{{ res.snippet }}</div>
+<div class="col-xs-6" *ngIf="showShuffled">
+    <ng-container>
+      <div class="results-container">
+        <app-text-list [data]="shuffledSearchResults"></app-text-list>
+      </div>
+    </ng-container>
   </div>
+  <div class="col-xs-6" *ngIf="(isMobile && !showShuffled)">
+    
+    <ng-container>
+      <div class="results-container">
+        <app-text-list [data]="unshuffledSearchResults"></app-text-list>
+      </div>
+    </ng-container>
+  </div>
+
+<div class="results-container">
+    <app-text-row *ngFor="let index of data" [data]="index"></app-text-row>
 </div>
